Fixed C's type decs to float from int

+++ bubblesort.c
@@ -4,7 +4,7 @@ typedef int bool;
 #define true 1
 #define false 0
 
-int* bubblesort(int array[], int size) {
+float* bubblesort(float array[], int size) {
 	bool swapping = false;
 	int temp;
 	int i;
@@ -24,8 +24,8 @@ int* bubblesort(int array[], int size) {
 
 int main(int argc, char* argv[]) {
 	int i = 0, count = 13;;
-	int array[] = {1337, 1, -465, 3.141592653589793, 789, 69, 789, -132, 3.141592653589793, 465, 789, 0, 27};
-	int* sorted = bubblesort(array, count);
+	float array[] = {1337, 1, -465, 3.141592653589793, 789, 69, 789, -132, 3.141592653589793, 465, 789, 0, 27};
+	float* sorted = bubblesort(array, count);
 	for (i=0; i < count; printf("%d ", sorted[i]), i++);
 	return 0;
 }
